Add the prebulit directory
This is an attempt to make "prebuilt" binaries on each platform where
the build will be run. In fact, they will only be built just prior
to running the high-level build.

Also, at this point, the plan is to use the same library as the library
for all the `PrebuiltLibraries` in the build. Let's see if that works.